// src/types/groupBuySale.ts
export interface GroupBuySale {
    Vendor: string;
    "Vendor rating": string;
    Type: string;
    Compound: string;
    Dose: string;
    Unit: string;
    Format: string;
    Quantity: string;
    Price: string;
    "Ships from Country": string;
    "Shipping $": string;
    MOQ: string;
    Analysis: string;
    "Purity guarantee": string;
    "Mass guarantee": string;
    "Re-ship guarantee": string;
    Start: string;
    Close: string;
    "PepChat Link": string;
    "Discord Link": string;
    "Telegram Link": string;
    Notes: string;
}